requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
dotenv==0.9.9
elevenlabs==2.22.0
flask==3.1.2
google-auth==2.43.0
google-genai==1.49.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.11.1
markupsafe==3.0.3
multidict==6.7.0
openai==2.7.1
packaging==25.0
pillow==12.0.0
postgrest==2.24.0
propcache==0.4.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pyjwt==2.10.1
python-dotenv==1.2.1
realtime==2.24.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
storage3==2.24.0
strenum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
tenacity==9.1.2
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
websockets==15.0.1
werkzeug==3.1.3
yarl==1.22.0
